Update checkout v2 line item cost to include multi-year discount
